+++ src/loaders/graphql.js
@@ -1,14 +1,14 @@
 const express_graphql = require("express-graphql");
 const bodyParser = require("body-parser");
-const graphiqlShcema = require("../models/graphql/schema");
-const graphqlResolvcer = require("../models/graphql/resolver");
+const graphiqlShcema = require("../models/graphql/schema/schema");
+const graphqlResolvcer = require("../models/graphql/resolver/resolver");
 const isAuth = require("../api/middlewares/is-auth");
 const { config } = require("../config");
 module.exports = graphql = app => {
   app.use(bodyParser.json());
   app.use(isAuth);
   app.use(
-    "/graphql",
+    `${config.api}/graphql`,
     express_graphql({
       schema: graphiqlShcema,
       rootValue: graphqlResolvcer,
