+++ src/models/block.js
@@ -1,7 +1,7 @@
 const mongoose = require("mongoose");
 
 const Schema = mongoose.Schema;
-const schema = new Schema({
+const Block = new Schema({
   name: {
     type: String,
     require: true
@@ -14,7 +14,11 @@ const schema = new Schema({
       type: Schema.Types.ObjectId,
       ref: "Subscription"
     }
-  ]
+  ],
+  blockAdmin: {
+    type: Schema.Types.ObjectId,
+    ref: "User"
+  }
 });
 
-module.exports = mongoose.model("Block", schema);
+module.exports = mongoose.model("Block", Block);
